// WebApi/ClientApp/src/model/respone/category/IMonHoc.ts
export interface IMonHoc {
  id_mon: number;
  ky_hieu: string;
  ten_viet_tat: string;
  ten_mon: string;
  ten_tieng_anh: string;
  ten_khoa: string;
  ten_bo_mon: string;
  id_bm: number;
  so_hoc_trinh: number;
  ly_thuyet: number;
  thuc_hanh: number;
  tu_hoc: number;
  id_khoa: number;
  is_active: boolean;
}